removing local OS files from public repo

+++ src/LEDwall_print/font_5x7.h
@@ -4,6 +4,7 @@
 #include <Arduino.h>
 
 // Declare the font table so other files can use it
-extern const uint8_t font5x7[][5];
+// declared as a 2D array, extern to define 
+extern const int font5x7[][5];
 
 #endif
